requirements.txt
amqp
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.5.0.2
celery==4.1.1
chardet==2.3.0
cssselect==0.9.2
dj-database-url==0.4.1
Django==1.10.2
django-annoying==0.10.3
django-bower==5.2.0
django-celery-beat==1.0.1
django-common-helpers==0.9.0
django-cors-headers==1.2.2
djangorestframework==3.4.7
gunicorn==19.6.0
pytz==2016.7
requests==2.11.1
six==1.10.0
subprocess.run==0.0.8
urllib3==1.7.1
whitenoise==3.2.2
psycopg2==2.6.2
django-cors-middleware==1.3.1
django_chartit
haversine
